// src/sanity/types.ts
export type ShopFacetQueryResult = Array<{
  _id: string;
  name: string | null;
  price: number | null;
  inStock: boolean | null;
  vehicleFit: Array<string> | null;
  categorySlug: string | null;
  categoryName: string | null;
}>;